+++ packages/multieditor/src/index.ts
@@ -1,4 +1,8 @@
-import { MultiEditor } from './internal';
+import { MultiEditor } from './multieditor';
 
-export * from './internal';
+export * from './multieditor';
+export * from './item';
+export * from './validator';
+export * from './status';
+
 export default MultiEditor;
