--- packages/multieditor/src/internal.ts
@@ -1,13 +1,0 @@
-// This file exposes Agile Editor functions and types to the outside world
-// It also serves as a cyclic dependency workaround
-// https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de.
-
-// !! All internal Agile Editor modules must be imported from here!!
-
-export * from './multieditor';
-export * from './item';
-export * from './validator';
-export * from './validator/types/string.validator';
-export * from './validator/types/number.validator';
-export * from './status/index';
-export * from './status/status.observer';
