+++ packages/multieditor/src/item.ts
@@ -3,7 +3,9 @@ import {
   defineConfig,
   EnhancedState,
 } from '@agile-ts/core';
-import { MultiEditor, Validator, Status, ItemKey } from './internal';
+import { ItemKey, MultiEditor } from './multieditor';
+import { Status } from './status';
+import { Validator } from './validator';
 
 export class Item<DataType = any> extends EnhancedState<DataType> {
   public editor: () => MultiEditor<DataType>;
